Fixed #8.

+++ .screen/rm_invalid_session_dirs.sh
@@ -9,9 +9,11 @@ for session_dir in "$HOME/.screen/sessions/"*; do
         break
     fi
     session=${session_dir##*/}
-    if echo "$list" | grep -Fq "$session"; then
-        :
-    else
-        rm -r "$session_dir"
+    if echo "$session" | grep -Eq "\.$(uname -n)\$"; then
+        if echo "$list" | grep -Fq "$session"; then
+            :
+        else
+            rm -r "$session_dir"
+        fi
     fi
 done
